update to remove an bit of the flag Weverything

+++ src/ft_str/ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include <unistd.h>
 
-size_t		ft_strlen(const char *str)
+int ft_strlen(const char *str)
 {
 	int i;
